# STM supports scala.js