--- a/core/js/src/main/scala/zio/internal/Sync.scala
+++ b/core/js/src/main/scala/zio/internal/Sync.scala
@@ -1,0 +1,8 @@
+package zio.internal
+
+object Sync {
+  def apply[A](anyRef: AnyRef)(f: => A): A = {
+    assert(anyRef != null)
+    f
+  }
+}