--- a/core/js/src/main/scala/zio/internal/Sync.scala
+++ b/core/js/src/main/scala/zio/internal/Sync.scala
@@ -2,7 +2,7 @@ package zio.internal
 
 object Sync {
   def apply[A](anyRef: AnyRef)(f: => A): A = {
-    assert(anyRef != null)
+    val _ = anyRef
     f
   }
 }