--- a/core/jvm/src/main/scala/zio/internal/Sync.scala
+++ b/core/jvm/src/main/scala/zio/internal/Sync.scala
@@ -1,0 +1,5 @@
+package zio.internal
+
+object Sync {
+  def apply[A](anyRef: AnyRef)(f: => A): A = anyRef.synchronized { f }
+}