--- a/core/shared/src/main/scala/zio/internal/LockedRef.scala
+++ b/core/shared/src/main/scala/zio/internal/LockedRef.scala
@@ -1,0 +1,32 @@
+package zio.internal
+
+class LockedRef[A](var ref: A) {
+  var locked = false
+
+  def get: A = ref
+  def set(a: A): A = {
+    this.ref = a
+    this.ref
+  }
+
+  def exclusiveSet(a: A): A = {
+    lock()
+    set(a)
+    unlock()
+    a
+  }
+
+  def lock(): Unit = {
+    var loop = true
+    while (loop) {
+      if (!locked) {
+        locked = true
+        loop = false
+      }
+    }
+  }
+
+  def unlock(): Unit =
+    locked = false
+
+}