--- a/core/shared/src/main/scala/zio/stm/STM.scala
+++ b/core/shared/src/main/scala/zio/stm/STM.scala
@@ -320,9 +320,10 @@ object STM {
     final def prepareResetJournal(journal: Journal): () => Unit = {
       val saved = new MutableMap[TRef[_], Entry](journal.size)
 
-      journal.forEach { (key, value) =>
-        saved.put(key, value.copy())
-        ()
+      val it = journal.entrySet.iterator
+      while (it.hasNext) {
+        val entry = it.next
+        saved.put(entry.getKey, entry.getValue.copy())
       }
 
       () => { journal.clear(); journal.putAll(saved); () }
@@ -331,8 +332,10 @@ object STM {
     /**
      * Commits the journal.
      */
-    final def commitJournal(journal: Journal): Unit =
-      journal.forEach((_, value) => value.commit())
+    final def commitJournal(journal: Journal): Unit = {
+      val it = journal.entrySet.iterator
+      while (it.hasNext) it.next.getValue.commit()
+    }
 
     /**
      * Allocates memory for the journal, if it is null, otherwise just clears it.
@@ -393,8 +396,9 @@ object STM {
       val allTodos  = new MutableMap[TxnId, Todo](DefaultJournalSize)
       val emptyTodo = Map.empty[TxnId, Todo]
 
-      journal.forEach { (_, value) =>
-        val tref = value.tref
+      val it = journal.entrySet.iterator
+      while (it.hasNext) {
+        val tref = it.next.getValue.tref
         val todo = tref.todo
 
         var loop = true
@@ -413,8 +417,10 @@ object STM {
     /**
      * Executes the todos in the current thread, sequentially.
      */
-    final def execTodos(todos: MutableMap[TxnId, Todo]): Unit =
-      todos.forEach((_, value) => value.apply())
+    final def execTodos(todos: MutableMap[TxnId, Todo]): Unit = {
+      val it = todos.entrySet.iterator
+      while (it.hasNext) it.next.getValue.apply()
+    }
 
     /**
      * For the given transaction id, adds the specified todo effect to all
@@ -423,8 +429,9 @@ object STM {
     final def addTodo(txnId: TxnId, journal: Journal, todoEffect: Todo): Boolean = {
       var added = false
 
-      journal.forEach { (_, value) =>
-        val tref = value.tref
+      val it = journal.entrySet.iterator
+      while (it.hasNext) {
+        val tref = it.next.getValue.tref
 
         var loop = true
         while (loop) {
@@ -462,7 +469,11 @@ object STM {
 
       untracked.putAll(newJournal)
 
-      newJournal.forEach { (key, value) =>
+      val it = newJournal.entrySet.iterator
+      while (it.hasNext) {
+        val entry = it.next
+        val key   = entry.getKey
+        val value = entry.getValue
         if (oldJournal.containsKey(key)) {
           // We already tracked this one, remove it:
           untracked.remove(key)
@@ -473,7 +484,6 @@ object STM {
           // succeed.
           untracked.remove(key)
         }
-        ()
       }
 
       untracked