@ghostdogpr 2019-08-03
Great work! Could you move the STM tests from `jvm` to `shared`? That would make them run with ScalaJS and validate the compatibility.