@heraklos 2019-08-03
I got away with `CountDownLatch` by replacing it with `Promise`.
But then I got new errors like the following:
```sbt
[error] java.util.concurrent.ExecutionException: Boxed Error
```

I'm trying to fix this.