@ghostdogpr 2019-08-03
@heraklos we could use a `Promise[Nothing, Unit]` instead of `CountDownLatch` but then `latch.countDown()` would have to be `unsafeRun(latch.succeed(()))` so that it is "unsafely" executed.

For example:
```scala
  def e21 =
    unsafeRun {
      for {
        latch <- Promise.make[Nothing, Unit]
        done  <- Promise.make[Nothing, Unit]
        tvar1 <- TRef.makeCommit(0)
        tvar2 <- TRef.makeCommit("Failed!")
        fiber <- (STM.atomically {
                  for {
                    v1 <- tvar1.get
                    _  <- STM.succeedLazy(unsafeRun(latch.succeed(())))
                    _  <- STM.check(v1 > 42)
                    _  <- tvar2.set("Succeeded!")
                    v2 <- tvar2.get
                  } yield v2
                } <* done.succeed(())).fork
        _    <- latch.await
        old  <- tvar2.get.commit
        _    <- tvar1.set(43).commit
        _    <- done.await
        newV <- tvar2.get.commit
        join <- fiber.join
      } yield (old must_=== "Failed!") and (newV must_=== join)
    }
```