@ghostdogpr 2019-08-03
Oops, you found it already. Are you getting those errors only with ScalaJS?