@heraklos 2019-08-03
@ghostdogpr 
Thank you for helping me about `CountDownLatch` :)

And yes to your question.  The error pops out only when executing ScalaJS test.
I noticed that there is `[error] Caused by: java.lang.Error: Cannot block for result to be set in JavaScript` in errors shown in the stack trace, and this message seems to come from `OneShot.get` but I cannot see any connection between STMSpec and this class...