@jdegoes 2019-08-03
The whole STM test suite should run on Scala.js without issue. But `unsafeRunSync` cannot be used. 

Specs 2 supports returning Futures, so we need to `unsafeRunToFuture` and tap into the Specs 2 support. 

Does that make sense?