@heraklos 2019-08-03
Thank you for your guide!
I'll replace all the tests that fail due to `unsafeRunSync` with  `unsafeRunToFuture` and spec2 Future support 👍