@heraklos 2019-08-04
Thanks a lot 😆

But another flaky test popped up... I'm sure this is the same issue as https://github.com/zio/zio/issues/1023
```
[error]     x supervise fibers in supervised
[error]  TIMEOUT: 60 seconds (TestRuntime.scala:18)
```
Would you run tests again?