@ghostdogpr 2019-08-04
LGTM, @jdegoes you wanna have a final look?